first project for medalgim

+++ lesson1/project1.py
@@ -1,7 +1,7 @@
 badget=int(input("please enter your marketing budget. \nfacebook campaign = 100ILS per day."
              "\ninstagram campaign = 50ILS per day."))
-fac_camp=int(input("how long do you want your facebook campaign to go?"))
-inst_camp=int(input("how long do you want your instagram campaign to go?"))
+fac_camp=float(input("how long do you want your facebook campaign to go?"))
+inst_camp=float(input("how long do you want your instagram campaign to go?"))
 price=(inst_camp*50 + fac_camp*100)*1.17
 print("your total price + tax will be:" + str(price))
 if(badget>price):
